pydex-registration/requirements.txt
Jinja2==3.1.2
google-auth==2.11.0
requests==2.28.1
pydash